text2: replacing old with new text implementation
