+++ ide/diff/editor-plugin.js
@@ -2,7 +2,6 @@ import { arr, string } from "lively.lang";
 import EditorPlugin from "../editor-plugin.js";
 import { pt } from "lively.graphics";
 import { config } from "lively.morphic";
-import { TextStyleAttribute } from "lively.morphic/text/attribute.js";
 import { lessEqPosition } from "lively.morphic/text//position.js";
 
 import DiffNavigator from "./navigator.js";
@@ -43,12 +42,13 @@ export class DiffEditorPlugin extends EditorPlugin {
 
     this.tokens = tokens;
     this.patches = patches;
+    
+    let attributes = [];
+    for (let {token, start, end} of tokens)
+      if (tokens.type !== "default")
+        attributes.push({start, end}, this.theme.styleCached(token));
+    textMorph.setTextAttributesWithSortedRanges(attributes);
 
-    var styles = tokens.map(({type, start, end}) =>
-          tokens.type !== "default" &&
-            TextStyleAttribute.fromPositions(this.theme.styleCached(type),start, end))
-              .filter(Boolean);
-    textMorph.setSortedTextAttributes([textMorph.defaultTextStyleAttribute].concat(styles));
   }
 
   getCommands(other) { return other.concat(commands); }
