+++ ide/js/editor-plugin.js
@@ -1,5 +1,4 @@
 import { string } from "lively.lang";
-import { TextStyleAttribute } from "../../text/attribute.js";
 import { lessEqPosition } from "../../text/position.js";
 import JavaScriptTokenizer from "./highlighter.js";
 import JavaScriptChecker from "./checker.js";
@@ -59,19 +58,12 @@ export class JavaScriptEditorPlugin extends EditorPlugin {
     
     textMorph.fill = this.theme.background();
     
-    let tokens = this._tokens = this.highlighter.tokenize(textMorph.textString);
+    let tokens = this._tokens = this.highlighter.tokenize(textMorph.textString),
+        attributes = [];
 
-    if (textMorph.setSortedTextAttributes) { //FIXME text transition 2017-04-07
-      textMorph.setSortedTextAttributes(
-        [textMorph.defaultTextStyleAttribute].concat(tokens.map(({token, start, end}) =>
-          TextStyleAttribute.fromPositions(this.theme.styleCached(token), start, end))));
-    } else {
-      let attributes = [];
-      for (let {token, start, end} of tokens)
-        attributes.push({start, end}, this.theme.styleCached(token));
-      textMorph.setTextAttributesWithSortedRanges(attributes);
-    }
-
+    for (let {token, start, end} of tokens)
+      attributes.push({start, end}, this.theme.styleCached(token));
+    textMorph.setTextAttributesWithSortedRanges(attributes);
 
     if (this.checker)
       this.checker.onDocumentChange({}, textMorph, this);
