+++ ide/js/workspace.js
@@ -6,7 +6,7 @@ import EvalBackendChooser from "./eval-backend-ui.js";
 import InputLine from "../../text/input-line.js";
 import { connect, once, noUpdate } from "lively.bindings";
 import { resource } from "lively.resources";
-import { Text } from "../../text2/morph.js";
+import { Text } from "../../text/morph.js";
 
 export default class Workspace extends Window {
