+++ tests/text/undo-test.js
@@ -1,6 +1,6 @@
 /*global System, declare, it, xit, describe, xdescribe, beforeEach, afterEach, before, after*/
 import { promise } from "lively.lang";
-import { Text } from "../../text2/morph.js";
+import { Text } from "../../text/morph.js";
 import { Range } from "../../text/selection.js";
 import { expect, chai } from "mocha-es6";
 import { dummyFontMetric as fontMetric, expectSelection } from "../test-helpers.js";
