+++ tests/text2/document-test.js
@@ -1,6 +1,6 @@
 /*global System, declare, it, xit, describe, xdescribe, beforeEach, afterEach, before, after*/
 import { expect } from "mocha-es6";
-import Document, { Line } from "../../text2/document.js";
+import Document, { Line } from "../../text/document.js";
 import { arr } from "lively.lang";
 
 var opts = {maxLeafSize: 3, minLeafSize: 2, maxNodeSize: 5, minNodeSize: 2};
