+++ text/input-line.js
@@ -3,7 +3,7 @@
 import { obj, arr, num } from "lively.lang";
 import { Rectangle, Color, pt } from "lively.graphics";
 import { connect, signal, disconnect } from "lively.bindings"; // for makeInputLine
-import { Text } from "../text2/morph.js"
+import { Text } from "../text/morph.js"
 import { Range } from "./range.js";
 
 export default class InputLine extends Text {
