+++ text/renderer.js
